feat: optimizes conditional logic performance
