+++ src/admin/components/forms/withCondition/index.tsx
@@ -1,12 +1,9 @@
 'use client';
 
-import React, { useEffect } from 'react';
+import React from 'react';
 import { FieldBase } from '../../../../fields/config/types';
-import { useAllFormFields } from '../Form/context';
-import getSiblingData from '../Form/getSiblingData';
-import reduceFieldsToValues from '../Form/reduceFieldsToValues';
-import { useDocumentInfo } from '../../utilities/DocumentInfo';
-import { useAuth } from '../../utilities/Auth';
+import { WatchCondition } from './WatchCondition';
+import { useFormFields } from '../Form/context';
 
 const withCondition = <P extends Record<string, unknown>>(Field: React.ComponentType<P>): React.FC<P> => {
   const CheckForCondition: React.FC<P> = (props) => {
@@ -26,7 +23,7 @@ const withCondition = <P extends Record<string, unknown>>(Field: React.Component
   const WithCondition: React.FC<P> = (props) => {
     const {
       name,
-      path: pathFromProps,
+      path,
       admin: {
         condition,
       } = {},
@@ -34,41 +31,28 @@ const withCondition = <P extends Record<string, unknown>>(Field: React.Component
       path?: string
     };
 
-    const path = typeof pathFromProps === 'string' ? pathFromProps : name;
+    const passesCondition = useFormFields(([fields]) => fields[path]?.passesCondition);
 
-    const { user } = useAuth();
-    const [fields, dispatchFields] = useAllFormFields();
-    const { id } = useDocumentInfo();
-
-    const data = reduceFieldsToValues(fields, true);
-    const siblingData = getSiblingData(fields, path);
-
-    // Manually provide ID to `data`
-    data.id = id;
-
-    const hasCondition = Boolean(condition);
-    const currentlyPassesCondition = hasCondition ? condition(data, siblingData, { user }) : true;
-    const field = fields[path];
-    const existingConditionPasses = field?.passesCondition;
-
-
-    useEffect(() => {
-      if (hasCondition) {
-        if (!existingConditionPasses && currentlyPassesCondition) {
-          dispatchFields({ type: 'MODIFY_CONDITION', path, result: true, user });
-        }
-
-        if (!currentlyPassesCondition && (existingConditionPasses || typeof existingConditionPasses === 'undefined')) {
-          dispatchFields({ type: 'MODIFY_CONDITION', path, result: false, user });
-        }
-      }
-    }, [currentlyPassesCondition, existingConditionPasses, dispatchFields, path, hasCondition, user]);
-
-    if (currentlyPassesCondition) {
-      return <Field {...props} />;
+    if (passesCondition) {
+      return (
+        <React.Fragment>
+          <WatchCondition
+            path={path}
+            name={name}
+            condition={condition}
+          />
+          <Field {...props} />
+        </React.Fragment>
+      );
     }
 
-    return null;
+    return (
+      <WatchCondition
+        path={path}
+        name={name}
+        condition={condition}
+      />
+    );
   };
 
   return CheckForCondition;
